--- screens/Home.js
@@ -1,12 +1,0 @@
-export const HomeScreen = ({ navigation }) => {
-  console.log("fdsfdsf")
-  return (
-    <Button
-      title="Go to Jane's profile"
-      onPress={() => navigation.navigate("Profile", { name: "Jane" })}
-    />
-  )
-}
-// export const ProfileScreen = ({ navigation, route }) => {
-//   return <Text>This is {route.params.name}'s profile</Text>
-// }
